Separate values from leafnodes
This commit removes the `value` field from `LeafNode`, allowing
values to be handled separately in the underlying data store. This
allows very efficent `get` operations, since the tree no longer
needs to be traversed in order to find the value of a given key.
The tree is still traversed if a proof is requested.