Support no_std (#87)
* Feature gate metrics

* Allow no_std.

This commit allows the JMT to be used in no_std environments. The trie
still requires alloc.

To implement this functionality, two non-trivial changes are made.
1. Replace all custom serialization with Borsh. **This is a breaking
   change which alters the state roots computed by the JMT**.
2. Remove support for multithreaded testing from the provided mock tree
   store. This functionality was unused, but if desired it can be easily
   re-implemented behind a feature-gate.

In addition, there are numerous minor changes to imports. Finally, the
tries built-in metrics are gated behind the "metrics" feature.

* Remove unused byteorder dep

* Bug fix: build jmt with std

A bug in the previous commit disabled all code when building with std.
This commit removes the erroneous #![cfg(...)] and fixes some imports
in the std version which caused builds to fail when the bug was fixed.

* Remove unused NodeDecodeError enum

* Add test vectors

* Switch MockTreeStore to parking lot

Remove test vector dep on std

* Remove deprecated encode/decode methods

* Add code to generate test vectors

* Bug fix: use alloc string/vec for test vectors

* Add ci for no_std, all features