jmt: implement update proofs (#97)
# Description
This PR allows the generation of merkle proofs in case of tree updates.
This PR is finally ready for review.

## Purpose
This PR aims to solve the issue [#246](https://github.com/Sovereign-Labs/sovereign-sdk/issues/246) of the Sovereign lab sdk repo 

## Methods added
The `verify_update` method have been added to the SparseMerkleProof, this method verifies a Merkle proof for a tree update.
The `put_value_set(s)_with_proof` methods have been added to the JellyfishMerkleTree struct. These methods update the tree and produce a set of Merkle proofs associated with the tree.

## Changes from existing implementation
- The `get_child_with_siblings` function now takes an object that implements the tree reader trait as an argument. This is due to deletion proofs that need to specify the sibling nodes' types (to know if we need to perform coalescing or not).
- The `TreeCache` wrapper object now implements the `TreeReader` trait. We need that to correctly build the sibling list in the `put_value_set(s)_with_proof` methods.
- I renamed `get_child_with_siblings` to `get_only_child_with_siblings`: this function did not have the same behavior as the `child` method of an internal node: when the nibble parameter points to an empty spot, `get_child_with_siblings` used to return the closest only child to that spot, while the `child` method would return none. Hence, I created a `get_child_with_siblings` reproducing the behavior of `child` and renamed the former method to `get_only_child_with_siblings`.

## Testing
A file `update_proofs.rs` have been added to the test set. This file reproduces most the tests from `jellyfish_merkle.rs` and checks that the updates have been well performed. This file contains proptests.